MOVEONLY: Move script/compressor out of script and put CTxOutCompressor (from core) with it
