+++ src/coins.h
@@ -6,8 +6,7 @@
 #ifndef BITCOIN_COINS_H
 #define BITCOIN_COINS_H
 
-#include "core.h" // Only for CTxOutCompressor
-#include "core/transaction.h"
+#include "compressor.h"
 #include "serialize.h"
 #include "uint256.h"
 #include "undo.h"
